% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{modArchive.info}
\alias{modArchive.info}
\alias{modArchive.download}
\alias{modArchive.search.mod}
\alias{modArchive.request.count}
\alias{modArchive.max.requests}
\alias{modArchive.view.by}
\alias{modArchive.search.genre}
\alias{modArchive.search.artist}
\alias{modArchive.search.hash}
\alias{modArchive.random.pick}
\alias{modLand.search.mod}
\alias{modLand.download.mod}
\title{Deprecated function migrated to \code{openmpt}}
\usage{
modArchive.info(...)

modArchive.download(...)

modArchive.search.mod(...)

modArchive.request.count(...)

modArchive.max.requests(...)

modArchive.view.by(...)

modArchive.search.genre(...)

modArchive.search.artist(...)

modArchive.search.hash(...)

modArchive.random.pick(...)

modLand.search.mod(...)

modLand.download.mod(...)
}
\arguments{
\item{...}{Ignored}
}
\value{
Returns \code{NULL}.
}
\description{
Functions listed on this page are deprecated and have migrated to the
\href{https://pepijn-devries.github.io/openmpt/articles/modarchive.html}{openmpt} package.
}
