% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06PTModule.r
\name{trackerFlag}
\alias{trackerFlag}
\alias{trackerFlag,PTModule-method}
\alias{trackerFlag<-}
\alias{trackerFlag<-,PTModule-method}
\title{Tracker flag indicating version compatibility}
\usage{
\S4method{trackerFlag}{PTModule}(x)

\S4method{trackerFlag}{PTModule}(x) <- value
}
\arguments{
\item{x}{A \code{\link{PTModule}} object for which the flag needs to
returned or replaced.}

\item{value}{A \code{character} string representing the tracker flag with which
that of object \code{x} needs to be replaced with. Should either be "\code{M.K.}"
or "\code{M!K!}". Note that if a current flag "\code{M!K!}" is
replaced by "\code{M.K.}", \code{\link{PTPattern}}s may get lost as the
latter supports less patterns.}
}
\value{
For \code{trackerFlag}, the tracker flag of object \code{x} is returned.

For \code{trackerFlag<-}, a copy of object \code{x} with an updated tracker
flag is returned.
}
\description{
Method to obtain a tracker flag, which indicates the version compatibility
of a ProTracker module (\code{\link{PTModule}} object).
}
\details{
ProTrackR supports two tracker flags: "\code{M.K.}" and "\code{M!K!}". M.K.
are presumably the initials of programmers Mahony and Kaktus, unfortunately
documentation on this matter is ambiguous. In any case, modules with the
flag "\code{M.K.}" can hold up to 64 patterns, whereas modules with the flag
"\code{M!K!}" can hold up to 100 patterns. Use this method to obtain or
replace the tracker flag of a \code{\link{PTModule}}.
}
\examples{
data("mod.intro")

## the current trackerFlag of mod.intro is "M.K.",
## meaning that it can hold a maximum of 64 patterns:
trackerFlag(mod.intro)

patternOrder(mod.intro, full = TRUE) <- 0:63

## If we upgrade the trackerFlag of mod.intro to "M!K!"
## it can hold a maximum of 100 patterns!:
trackerFlag(mod.intro) <- "M!K!"

patternOrder(mod.intro, full = TRUE) <- 0:99

## Now let's do something dangerous:
## current flag is "M!K!", by setting it
## back to "M.K.", patterns 65:100 are lost...
trackerFlag(mod.intro) <- "M.K."

}
\seealso{
Other module.operations: 
\code{\link{PTModule-class}},
\code{\link{appendPattern}()},
\code{\link{clearSamples}()},
\code{\link{clearSong}()},
\code{\link{deletePattern}()},
\code{\link{fix.PTModule}()},
\code{\link{modToWave}()},
\code{\link{moduleSize}()},
\code{\link{patternLength}()},
\code{\link{patternOrderLength}()},
\code{\link{patternOrder}()},
\code{\link{playMod}()},
\code{\link{playingtable}()},
\code{\link{rawToPTModule}()},
\code{\link{read.module}()},
\code{\link{write.module}()}
}
\author{
Pepijn de Vries
}
\concept{module.operations}
