% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01supporting_functions.r
\name{sampleRate}
\alias{sampleRate}
\alias{noteToSampleRate}
\alias{periodToSampleRate}
\title{Calculate the sample rate for a note or period value}
\usage{
noteToSampleRate(note = "C-3", finetune = 0, video = c("PAL", "NTSC"))

periodToSampleRate(period, video = c("PAL", "NTSC"))
}
\arguments{
\item{note}{A \code{character} string representing a note for which the sample
rate is to be calculated.}

\item{finetune}{An \code{integer} value ranging from -8 up to 7. A value used to
tune an audio sample.}

\item{video}{The video mode used to calculate the sample rate. A \code{character}
string that can have either the value '\href{https://en.wikipedia.org/wiki/PAL}{PAL}'
or '\href{https://en.wikipedia.org/wiki/NTSC}{NTSC}'. PAL is used by default.}

\item{period}{A ProTracker \code{integer} value of a period value for which the sample rate
is to be calculated.}
}
\value{
Returns the sample rate in samples per seconds.
}
\description{
Calculate the sample rate for a note or a ProTracker period value.
}
\details{
The timing on a Commodore Amiga depends on the video mode, which could be
either '\href{https://en.wikipedia.org/wiki/PAL}{PAL}'
or '\href{https://en.wikipedia.org/wiki/NTSC}{NTSC}'. Therefore sample
rates also depend on these modes. As the PAL is mostly used in Europe, and
the Amiga was most popular in Europe, PAL is used by default.
}
\examples{
## calculate the sample rate for a ProTracker period value of 200
periodToSampleRate(200)

## calculate the sample rate for a sample at note 'A-3'
noteToSampleRate("A-3")

## note that the NTSC video system gives a slightly different rate:
noteToSampleRate("A-3", video = "NTSC")

## fine tuning a sample will also give a slightly different rate:
noteToSampleRate("A-3", finetune = -1)

}
\seealso{
Other character.operations: 
\code{\link{as.character}()},
\code{\link{name}},
\code{\link{periodToChar}()},
\code{\link{rawToCharNull}()}

Other period.operations: 
\code{\link{noteToPeriod}()},
\code{\link{note}()},
\code{\link{octave}()},
\code{\link{periodToChar}()},
\code{\link{period_table}}

Other sample.rate.operations: 
\code{\link{playSample}()}

Other note.and.octave.operations: 
\code{\link{noteToPeriod}()},
\code{\link{noteUp}()},
\code{\link{note}()},
\code{\link{octave}()},
\code{\link{periodToChar}()}
}
\author{
Pepijn de Vries
}
\concept{character.operations}
\concept{note.and.octave.operations}
\concept{period.operations}
\concept{sample.rate.operations}
